Code - Web Application/ai_gcp/requirements.txt
h5py==3.10.0
tensorboard==2.15.0
tensorflow==2.15.0
pandas==2.2.1
numpy==1.26.4
joblib==1.3.2
google-cloud-storage
google-cloud-pubsub
scikit-learn==1.2.2
Flask
Pillow
gunicorn==22.0.0